ci: remove publish.yml GitHub Actions workflow
Deleted the `publish.yml` file, which defined a GitHub Actions workflow for publishing NuGet packages. The workflow was triggered on `push` events for tags matching the `v*` pattern and included steps for building, packing, and publishing the `CrossPrimitives` project to GitHub Packages and NuGet.org.

This change removes the automated publishing process from the repository.